requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-filter==21.1
pandas==1.4.2